backend/pyproject.toml
[project]
name = "vaultx-backend"
version = "0.1.0"
description = "VaultX backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt<4.1",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.121.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.44",
]
